fix: fix inventory model location
