+++ apps/web/src/schema/_common/model.ts
@@ -9,7 +9,7 @@ export interface TimeStamp {
 
 export interface ListResponse<T> {
     results: T[];
-    total_count?: number;
+    total_count: number;
 }
 
 export interface AnalyzeResponse<T> {
