+++ apps/web/src/services/asset-inventory/components/CollectorDetailPluginSummaryCards.vue
@@ -6,7 +6,8 @@ import { PCopyButton, PFieldTitle } from '@spaceone/design-system';
 
 import { iso8601Formatter } from '@cloudforet/core-lib/index';
 
-import type { CollectorModel, JobModel } from '@/schema/inventory/collector/model';
+import type { CollectorModel } from '@/schema/inventory/collector/model';
+import type { JobModel } from '@/schema/inventory/job/model';
 import { store } from '@/store';
 
 import RecentCollectorJobList from '@/services/asset-inventory/components/RecentCollectorJobList.vue';
