+++ apps/web/src/services/asset-inventory/stores/collector-form-store.ts
@@ -5,8 +5,9 @@ import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 
 import type {
     CollectorModel, CollectorOptions,
-    RepositoryPluginModel,
+
 } from '@/schema/inventory/collector/model';
+import type { PluginModel } from '@/schema/repository/plugin/model';
 import { store } from '@/store';
 import { i18n } from '@/translations';
 
@@ -35,7 +36,7 @@ export type ServiceAccountFilterOption = 'include'|'exclude';
 export const useCollectorFormStore = defineStore('collector-form', {
     state: () => ({
         originCollector: null as CollectorModel|null, // data from inventory.collector.get api.
-        repositoryPlugin: null as RepositoryPluginModel|null, // data from repository.plugin.list api. it's used when creating collector.
+        repositoryPlugin: null as PluginModel|null, // data from repository.plugin.list api. it's used when creating collector.
         provider: null as string|null,
         tags: {} as Tag,
         name: '',
@@ -68,7 +69,7 @@ export const useCollectorFormStore = defineStore('collector-form', {
             this.originCollector = collector;
             await this.resetForm();
         },
-        setRepositoryPlugin(pluginInfo: RepositoryPluginModel|null) {
+        setRepositoryPlugin(pluginInfo: PluginModel|null) {
             this.repositoryPlugin = pluginInfo;
         },
         async resetForm() {
