+++ apps/web/src/services/asset-inventory/types/collector-main-page-type.ts
@@ -2,10 +2,12 @@ import type { ConsoleFilter } from '@cloudforet/core-lib/query/type';
 
 import type {
     CollectorPluginModel, Schedule,
-    JobStatus,
+
 } from '@/schema/inventory/collector/model';
+import type { JobStatus } from '@/schema/inventory/job/type';
 
 import type { RouteQueryString } from '@/lib/router-query-string';
+
