+++ apps/web/src/store/modules/reference/plugin/actions.ts
@@ -3,7 +3,7 @@ import type { Action } from 'vuex';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 
 
-import type { RepositoryPluginModel } from '@/schema/inventory/collector/model';
+import type { PluginModel } from '@/schema/repository/plugin/model';
 
 import { REFERENCE_LOAD_TTL } from '@/store/modules/reference/config';
 import type { PluginReferenceMap, PluginReferenceState } from '@/store/modules/reference/plugin/type';
@@ -43,7 +43,7 @@ export const load: Action<PluginReferenceState, any> = async ({ state, commit },
                 repository_id: repoInfo.repository_id,
             }, { timeout: 3000 });
 
-            pluginResponse.results.forEach((pluginInfo: RepositoryPluginModel): void => {
+            pluginResponse.results.forEach((pluginInfo: PluginModel): void => {
                 plugins[pluginInfo.plugin_id] = {
                     key: pluginInfo.plugin_id,
                     label: pluginInfo.name,
